The renaming that Git forgot

+++ src/main/kotlin/com/scientianova/versatile/machines/gui/textures/IDrawable.kt
@@ -1,10 +1,10 @@
 package com.scientianova.versatile.machines.gui.textures
 
 interface IDrawable {
-    val uStart: Double
-    val vStart: Double
-    val uEnd: Double
-    val vEnd: Double
+    val uStart: Float
+    val vStart: Float
+    val uEnd: Float
+    val vEnd: Float
 
-    fun draw(x: Int, y: Int, width: Int, height: Int, uStart: Double = this.uStart, vStart: Double = this.vStart, uEnd: Double = this.uEnd, vEnd: Double = this.vEnd)
+    fun draw(x: Int, y: Int, width: Int, height: Int, uStart: Float = this.uStart, vStart: Float = this.vStart, uEnd: Float = this.uEnd, vEnd: Float = this.vEnd)
 }
